% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_fun.R
\name{AllGatedScores}
\alias{AllGatedScores}
\title{All Scoring Function}
\usage{
AllGatedScores(df, colname)
}
\arguments{
\item{df}{a dataframe for scoring}

\item{colname}{column name for gated score}
}
\value{
a dataframe with gated scores from z scores
}
\description{
This function allows you to calculated the 5 level gated scores for all reports
}
