% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_fun.R
\name{pick_vars}
\alias{pick_vars}
\title{Pick Var Function}
\usage{
pick_vars(df, var_names)
}
\arguments{
\item{df}{a dataframe for scoring}

\item{var_names}{a vector of variables used for each domain}
}
\value{
a dataframe with selected columns
}
\description{
This function allows you to select the listed columns from a DF
}
