% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_fun.R
\name{scores
This function allows you to calculated z-scores and 5 level gated scores}
\alias{scores
This function allows you to calculated z-scores and 5 level gated scores}
\alias{scores}
\title{Scoring Function}
\usage{
scores(df, var_names)
}
\arguments{
\item{df}{a dataframe for scoring}

\item{var_names}{a vector of variables used for each domain}
}
\value{
a dataframe with total score for the domain, zscores and gated scores from the total score
}
\description{
Scoring Function
}
\keyword{,}
\keyword{gated}
\keyword{scores}
\keyword{z-scores}
